split home page
