+++ components/page-components/Home/index.jsx
@@ -1,0 +1,19 @@
+import Head from 'next/head';
+
+function Home() {
+    return (
+        <div>
+            <Head>
+                <title>Book Teaser Website</title>
+                <meta
+                    name='description'
+                    content='For those who likes to read'
+                />
+            </Head>
+
+            <main>hi</main>
+        </div>
+    );
+}
+
+export default Home;
