+++ pages/index.js
@@ -1,19 +1,2 @@
-import Head from 'next/head';
-
-function Home() {
-    return (
-        <div>
-            <Head>
-                <title>Book Teaser Website</title>
-                <meta
-                    name='description'
-                    content='For those who likes to read'
-                />
-            </Head>
-
-            <main>hi</main>
-        </div>
-    );
-}
-
+import Home from '../components/page-components/Home';
 export default Home;
